backend/requirements.txt
# FastAPI + Dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# Database
sqlmodel>=0.0.22
aiosqlite>=0.20.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Email
resend>=2.5.0

# Utilities
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
python-dotenv>=1.0.0

# Development
pytest>=8.3.0
httpx>=0.28.0
